chore: hacer capa  de datos entera de tarea y parte de operacion

+++ shared/src/commonMain/kotlin/dev/wdona/burnt_out/domain/model/Tarea.kt
@@ -3,5 +3,5 @@ package dev.wdona.burnt_out.shared.domain
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Tarea(val idTarea: Long, val titulo: String, val descripcion: String?, val estado: String, val idTableroPerteneciente: Long, val idUsuarioAsignado: Long, val idSubtareas: List<Int>?)
+data class Tarea(val idTarea: Long, val titulo: String, val descripcion: String?, val estado: String, val idTableroPerteneciente: Long, val idUsuarioAsignado: Long, val idSubtareas: List<Long>?)
